backend/requirements.txt
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
python-dotenv
requests
gunicorn
psycopg2-binary
datetime
stripe
Flask-Caching
redis
prometheus-flask-exporter==0.23.0
pandas
pyarrow